Fix files handling of revoked user thx randomshell
When revoking a user, if a new user is then created with the same name, it will generate a bad config that cannot used. This is due to the old config not being deleted, or overwritten. Instead, the new config is appended to the end of the file.

It would make more sense to either overwrite the file entirely, prompt the user if it should be replaced during the creation of the new user, or even prompt for the config to be deleted when revoking the user.

Simply deleting the offending config before re-creating the user will resolve this issue.

this fix it now